+++ dubbo-remoting/dubbo-remoting-http12/src/main/java/org/apache/dubbo/remoting/http12/h1/Http1ServerChannelObserver.java
@@ -17,91 +17,17 @@
 package org.apache.dubbo.remoting.http12.h1;
 
 import org.apache.dubbo.remoting.http12.AbstractServerHttpChannelObserver;
-import org.apache.dubbo.remoting.http12.ErrorResponse;
 import org.apache.dubbo.remoting.http12.HttpChannel;
 import org.apache.dubbo.remoting.http12.HttpChannelObserver;
-import org.apache.dubbo.remoting.http12.HttpHeaderNames;
 import org.apache.dubbo.remoting.http12.HttpHeaders;
 import org.apache.dubbo.remoting.http12.HttpMetadata;
 import org.apache.dubbo.remoting.http12.HttpOutputMessage;
-import org.apache.dubbo.remoting.http12.HttpResult;
-import org.apache.dubbo.remoting.http12.HttpStatus;
-import org.apache.dubbo.remoting.http12.exception.EncodeException;
-import org.apache.dubbo.remoting.http12.exception.HttpResultPayloadException;
-import org.apache.dubbo.remoting.http12.exception.HttpStatusException;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import io.netty.buffer.ByteBufOutputStream;
 
 public class Http1ServerChannelObserver extends AbstractServerHttpChannelObserver
         implements HttpChannelObserver<Object> {
 
     public Http1ServerChannelObserver(HttpChannel httpChannel) {
         super(httpChannel);
-    }
-
-    @Override
-    public void onNext(Object data) {
-        if (supportChunk()) {
-            super.onNext(data);
-            return;
-        }
-        try {
-            String status = HttpStatus.OK.getStatusString();
-            Map<String, List<String>> additionalHeaders = null;
-            if (data instanceof HttpResult) {
-                HttpResult<?> result = (HttpResult<?>) data;
-                data = result.getBody();
-                status = String.valueOf(result.getStatus());
-                additionalHeaders = result.getHeaders();
-            }
-            HttpOutputMessage outputMessage = encodeHttpOutputMessage(data);
-            preOutputMessage(outputMessage);
-            getResponseEncoder().encode(outputMessage.getBody(), data);
-            if (!headerSent) {
-                doSendHeaders(status, buildAdditionalHeaders(outputMessage, additionalHeaders));
-            }
-            getHttpChannel().writeMessage(outputMessage);
-            postOutputMessage(outputMessage);
-        } catch (Throwable e) {
-            onError(e);
-        }
-    }
-
-    @Override
-    public void onError(Throwable throwable) {
-        if (supportChunk()) {
-            super.onError(throwable);
-            return;
-        }
-        if (throwable instanceof HttpResultPayloadException) {
-            onNext(((HttpResultPayloadException) throwable).getResult());
-            return;
-        }
-        int httpStatusCode = HttpStatus.INTERNAL_SERVER_ERROR.getCode();
-        if (throwable instanceof HttpStatusException) {
-            httpStatusCode = ((HttpStatusException) throwable).getStatusCode();
-        }
-        try {
-            ErrorResponse errorResponse = new ErrorResponse();
-            errorResponse.setStatus(String.valueOf(httpStatusCode));
-            errorResponse.setMessage(throwable.getMessage());
-            getErrorResponseCustomizer().accept(errorResponse, throwable);
-            HttpOutputMessage httpOutputMessage = encodeHttpOutputMessage(errorResponse);
-            getResponseEncoder().encode(httpOutputMessage.getBody(), errorResponse);
-            if (!headerSent) {
-                doSendHeaders(String.valueOf(httpStatusCode), buildAdditionalHeaders(httpOutputMessage, null));
-            }
-            getHttpChannel().writeMessage(httpOutputMessage);
-        } catch (Throwable ex) {
-            throwable = new EncodeException(ex);
-        } finally {
-            doOnCompleted(throwable);
-        }
     }
 
     @Override
@@ -114,20 +40,5 @@ public class Http1ServerChannelObserver extends AbstractServerHttpChannelObserve
     protected void doOnCompleted(Throwable throwable) {
         super.doOnCompleted(throwable);
         this.getHttpChannel().writeMessage(HttpOutputMessage.EMPTY_MESSAGE);
-    }
-
-    private Map<String, List<String>> buildAdditionalHeaders(
-            HttpOutputMessage outputMessage, Map<String, List<String>> additionalHeaders) {
-        int contentLength = ((ByteBufOutputStream) outputMessage.getBody()).writtenBytes();
-        if (additionalHeaders == null) {
-            additionalHeaders = new HashMap<>();
-        }
-        additionalHeaders.put(
-                HttpHeaderNames.CONTENT_LENGTH.getName(), Collections.singletonList(String.valueOf(contentLength)));
-        return additionalHeaders;
-    }
-
-    protected boolean supportChunk() {
-        return false;
     }
 }
